Move functions for unprivileged access to API dir
The functions for unprivileged access are useful for user code that can
run in privileged mode.
